fix Print following pull request comments
