+++ package.json
@@ -27,8 +27,7 @@
   },
   "dependencies": {
     "@material-ui/core": "^4.10.1",
-    "@material-ui/icons": "^4.5.1",
-    "regenerator-runtime": "^0.13.7"
+    "@material-ui/icons": "^4.5.1"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
