+++ src/Print/Button/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Button from "./Button";
+
+export default Button;
