+++ src/Print/Printing/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Printing from "./Printing";
+
+export default Printing;
