+++ src/Search/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Search from "./Search";
+
+export default Search;
