+++ src/style.css
@@ -1,7 +1,3 @@
 .o2xp-to-print-columns{
     color: rgba(0, 0, 0, 0.54);
 }
-
-.MuiDialog-paperWidthXl{
-    width: 600px;
-}
